Create MenuItem class
Created menuitem class for testing (it was put in mistakenly last push)
Deleted the class (it served it's purpose)
So basically a few tasks kind of got rolled into one.
When play is pressed, a start game and quit game button populate, when clicked, a LOG is printed to show button clicked.